Time: 46 ms (12.55%) | Memory: 55.9 MB (37.12%) - LeetSync

+++ 1675-magnetic-force-between-two-balls/magnetic-force-between-two-balls.java
@@ -1,5 +1,5 @@
 class Solution {
-    private boolean isPossible(int[]position,int m,long dis){
+    private boolean isPossible(int[]position,int m,int dis){
         int last = 0;
         for(int i=1;i<position.length;i++){
             if(Math.abs(position[last]-position[i]) >= dis) {
@@ -12,14 +12,14 @@ class Solution {
     }
     public int maxDistance(int[] position, int m) {
         Arrays.sort(position);
-        long st = 1 , end = position[position.length-1]-position[0];
+        int st = 1 , end = position[position.length-1]-position[0];
         while(st <= end){
-            long dis = st+(end-st)/2;
+            int dis = st+(end-st)/2;
             if(isPossible(position,m,dis)){
                 st = dis+1;
             }
             else end = dis-1;
         }
-        return (int)end;
+        return end;
     }
 }
